nil protect when destroy called for tech metadata

+++ app/jobs/replace_master_files.rb
@@ -32,7 +32,7 @@ class ReplaceMasterFiles < BaseJob
 
          # update MF attributes, replace tech metadata, re-publish and archive
          curr_mf.update(filesize: fs, md5: md5)
-         curr_mf.image_tech_meta.destroy
+         curr_mf.image_tech_meta.destroy if !curr_mf.image_tech_meta.nil?
          CreateImageTechnicalMetadata.exec_now({master_file: curr_mf, source: mf_path}, self)
          PublishToIiif.exec_now({source: mf_path, master_file_id: curr_mf.id, overwrite: true}, self)
